refactor module_utils
