+++ plugins/module_utils/ciscosmb_canonical_map.py
