+++ plugins/modules/ciscosmb_command.py
@@ -111,8 +111,8 @@ failed_conditions:
 import time
 
 from ansible_collections.ansible.netcommon.plugins.module_utils.network.common.parsing import Conditional
-from ansible_collections.qaxi.ciscosmb.plugins.module_utils.network.ciscosmb.ciscosmb import run_commands
-from ansible_collections.qaxi.ciscosmb.plugins.module_utils.network.ciscosmb.ciscosmb import ciscosmb_argument_spec
+from ansible_collections.qaxi.ciscosmb.plugins.module_utils.ciscosmb import run_commands
+from ansible_collections.qaxi.ciscosmb.plugins.module_utils.ciscosmb import ciscosmb_argument_spec
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.six import string_types
